requirements.txt
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.23.0,<1.0
SQLAlchemy>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9
openai>=1.12.0
google-genai>=0.1.0
